Added checking on collisions with obstacles. Now it is impossible to pass through obstacles

+++ data/src/jvmMain/kotlin/mapper/OffsetMapper.kt
@@ -5,6 +5,6 @@ import model.Point
 
 internal class OffsetMapper {
     fun toPoint(offset: Offset): Point {
-        return Point(offset.x.toDouble(), offset.y)
+        return Point(offset.x, offset.y)
     }
 }
